ai-rag-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
google-generativeai==0.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
huggingface-hub<0.24.0
